Python/data-connection/BLOB Python Connector/requirements.txt
azure-storage-blob==12.0.0
azure-storage==0.36.0
